refactor FollowButton with more general component
